backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Shared Stickers backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.47",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.2.1",
    "boto3>=1.34.0",
]